+++ layers/silver/sku_id/sku_cm2_inputs_daily/v1/migrations/create_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS {target_table} (
     dt TIMESTAMP COMMENT 'Начало даты выполнения расчёта (00:00:00 Asia/Tashkent); часть уникального ключа dt, sku_id',
-    sku_id INT COMMENT 'Идентификатор SKU; часть уникального ключа dt, sku_id',
+    sku_id BIGINT COMMENT 'Идентификатор SKU; часть уникального ключа dt, sku_id',
     product_id INT COMMENT 'Идентификатор товара, по которому Main CM2 и PDP CM2 агрегируют SKU-значения в Gold',
     dimensional_group STRING COMMENT 'Габаритная группа SKU: SMALL, MEDIUM или LARGE; NULL источника заменяется на SMALL',
     sell_price_uzs DOUBLE COMMENT 'Историческая EOD sell price SKU в UZS за календарный день перед dt либо NULL',
